Fixed the logic for the validation.
