+++ Services/UI/OJS.Services.Ui.Models/Contests/ContestParticipationServiceModel.cs
@@ -29,6 +29,7 @@ public class ContestParticipationServiceModel : IMapExplicitly
 
     public void RegisterMappings(IProfileExpression configuration)
         => configuration.CreateMap<Participant, ContestParticipationServiceModel>()
+            .ForMember(d => d.Contest, opt => opt.Ignore())
             .ForMember(d => d.EndDateTimeForParticipantOrContest, opt => opt.Ignore())
             .ForAllOtherMembers(opt => opt.Ignore());
 }
